// modules/frontend/src/modules/core/services/generated/model/anamnesisCooperation.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * HISTORIA CLINICA ORTODONCIA
 * OpenAPI spec version: v1
 */

export type AnamnesisCooperation = typeof AnamnesisCooperation[keyof typeof AnamnesisCooperation];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AnamnesisCooperation = {
  HIGH: 'HIGH',
  MEDIUM: 'MEDIUM',
  LOW: 'LOW',
} as const;