// modules/frontend/src/modules/core/services/generated/model/angleRelationAngleClass.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * HISTORIA CLINICA ORTODONCIA
 * OpenAPI spec version: v1
 */

export type AngleRelationAngleClass = typeof AngleRelationAngleClass[keyof typeof AngleRelationAngleClass];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const AngleRelationAngleClass = {
  CLASS_I: 'CLASS_I',
  CLASS_II: 'CLASS_II',
  CLASS_III: 'CLASS_III',
} as const;