// modules/frontend/src/modules/core/services/generated/model/createRecordRequestDocumentType.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * Tu portal al conocimiento y al éxito profesional.
 * OpenAPI spec version: v1
 */

export type CreateRecordRequestDocumentType = typeof CreateRecordRequestDocumentType[keyof typeof CreateRecordRequestDocumentType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateRecordRequestDocumentType = {
  DNI: 'DNI',
  FOREIGNER_CARD: 'FOREIGNER_CARD',
  PASSPORT: 'PASSPORT',
} as const;