// modules/frontend/src/modules/core/services/generated/model/createRecordRequestPatientSex.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * Tu portal al conocimiento y al éxito profesional.
 * OpenAPI spec version: v1
 */

export type CreateRecordRequestPatientSex = typeof CreateRecordRequestPatientSex[keyof typeof CreateRecordRequestPatientSex];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateRecordRequestPatientSex = {
  FEMALE: 'FEMALE',
  MALE: 'MALE',
} as const;