// modules/frontend/src/modules/core/services/generated/model/facialAnalysisChinNeckAngle.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * HISTORIA CLINICA ORTODONCIA
 * OpenAPI spec version: v1
 */

export type FacialAnalysisChinNeckAngle = typeof FacialAnalysisChinNeckAngle[keyof typeof FacialAnalysisChinNeckAngle];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FacialAnalysisChinNeckAngle = {
  NORMAL: 'NORMAL',
  OPEN: 'OPEN',
  CLOSED: 'CLOSED',
} as const;