// modules/frontend/src/modules/core/services/generated/model/facialAnalysisChinNeckLine.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * Tu portal al conocimiento y al éxito profesional.
 * OpenAPI spec version: v1
 */

export type FacialAnalysisChinNeckLine = typeof FacialAnalysisChinNeckLine[keyof typeof FacialAnalysisChinNeckLine];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FacialAnalysisChinNeckLine = {
  NORMAL: 'NORMAL',
  INCREASED: 'INCREASED',
  DECREASED: 'DECREASED',
} as const;