// modules/frontend/src/modules/core/services/generated/model/facialAnalysisConvexity.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * HISTORIA CLINICA ORTODONCIA
 * OpenAPI spec version: v1
 */

export type FacialAnalysisConvexity = typeof FacialAnalysisConvexity[keyof typeof FacialAnalysisConvexity];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FacialAnalysisConvexity = {
  STRAIGHT: 'STRAIGHT',
  CONVEX: 'CONVEX',
  CONCAVE: 'CONCAVE',
} as const;