// modules/frontend/src/modules/core/services/generated/model/facialAnalysisFacialPattern.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * HS FACOP API
 * HISTORIA CLINICA ORTODONCIA
 * OpenAPI spec version: v1
 */

export type FacialAnalysisFacialPattern = typeof FacialAnalysisFacialPattern[keyof typeof FacialAnalysisFacialPattern];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const FacialAnalysisFacialPattern = {
  PATTERN_I: 'PATTERN_I',
  PATTERN_II: 'PATTERN_II',
  PATTERN_III: 'PATTERN_III',
  SHORT_FACE: 'SHORT_FACE',
  LONG_FACE: 'LONG_FACE',
} as const;